+++ pyriemann/geometry/tangentspace.py
@@ -13,7 +13,7 @@ from ._check import check_function, check_matrix_pair
 from .base import ctranspose, expm, invsqrtm, logm, sqrtm, ddexpm, ddlogm
 
 
-def exp_map_euclid(X, Cref):
+def exp_map_euclid(X, Cref, **kwargs):
     r"""Project matrices back to manifold by Euclidean exponential map.
 
     The projection of a matrix :math:`\mathbf{X}` from tangent space
@@ -44,7 +44,7 @@ def exp_map_euclid(X, Cref):
     return X + Cref
 
 
-def exp_map_logchol(X, Cref):
+def exp_map_logchol(X, Cref, **kwargs):
     r"""Project matrices back to manifold by log-Cholesky exponential map.
 
     The projection of a matrix :math:`\mathbf{X}` from tangent space
@@ -99,7 +99,7 @@ def exp_map_logchol(X, Cref):
     return exp_map @ ctranspose(exp_map)
 
 
-def exp_map_logeuclid(X, Cref):
+def exp_map_logeuclid(X, Cref, **kwargs):
     r"""Project matrices back to manifold by log-Euclidean exponential map.
 
     The projection of a matrix :math:`\mathbf{X}` from tangent space
@@ -149,7 +149,7 @@ def exp_map_logeuclid(X, Cref):
     return expm(logm(Cref) + ddlogm(X, Cref))
 
 
-def exp_map_riemann(X, Cref, Cm12=False):
+def exp_map_riemann(X, Cref, *, Cm12=False):
     r"""Project matrices back to manifold by Riemannian exponential map.
 
     The projection of a matrix :math:`\mathbf{X}` from tangent space
@@ -202,7 +202,7 @@ def exp_map_riemann(X, Cref, Cm12=False):
     return C12 @ expm(X) @ C12
 
 
-def exp_map_wasserstein(X, Cref):
+def exp_map_wasserstein(X, Cref, **kwargs):
     r"""Project matrices back to manifold by Wasserstein exponential map.
 
     The projection of a matrix :math:`\mathbf{X}` from tangent space
@@ -257,7 +257,7 @@ exp_map_functions = {
 }
 
 
-def exp_map(X, Cref, *, metric="riemann"):
+def exp_map(X, Cref, *, metric="riemann", **kwargs):
     """Project matrices back to manifold by exponential map.
 
     Parameters
@@ -270,6 +270,10 @@ def exp_map(X, Cref, *, metric="riemann"):
         Metric used for exponential map, can be:
         "euclid", "logchol", "logeuclid", "riemann", "wasserstein",
         or a callable function.
+    **kwargs : dict
+        The keyword arguments passed to the sub function.
+
+        .. versionadded:: 0.12
 
     Returns
     -------
@@ -291,13 +295,13 @@ def exp_map(X, Cref, *, metric="riemann"):
     exp_map_wasserstein
     """
     exp_map_function = check_function(metric, exp_map_functions)
-    return exp_map_function(X, Cref)
+    return exp_map_function(X, Cref, **kwargs)
 
 
 ###############################################################################
 
 
-def log_map_euclid(X, Cref):
+def log_map_euclid(X, Cref, **kwargs):
     r"""Project matrices in tangent space by Euclidean logarithmic map.
 
     The projection of a matrix :math:`\mathbf{X}` from manifold
@@ -328,7 +332,7 @@ def log_map_euclid(X, Cref):
     return X - Cref
 
 
-def log_map_logchol(X, Cref):
+def log_map_logchol(X, Cref, **kwargs):
     r"""Project matrices in tangent space by log-Cholesky logarithmic map.
 
     The projection of a matrix :math:`\mathbf{X}` from SPD/HPD manifold
@@ -375,7 +379,7 @@ def log_map_logchol(X, Cref):
     return X_new
 
 
-def log_map_logeuclid(X, Cref):
+def log_map_logeuclid(X, Cref, **kwargs):
     r"""Project matrices in tangent space by log-Euclidean logarithmic map.
 
     The projection of a matrix :math:`\mathbf{X}` from SPD/HPD manifold
@@ -427,7 +431,7 @@ def log_map_logeuclid(X, Cref):
     return X_new
 
 
-def log_map_riemann(X, Cref, C12=False):
+def log_map_riemann(X, Cref, *, C12=False):
     r"""Project matrices in tangent space by Riemannian logarithmic map.
 
     The projection of a matrix :math:`\mathbf{X}` from SPD/HPD manifold
@@ -481,7 +485,7 @@ def log_map_riemann(X, Cref, C12=False):
     return X_new
 
 
-def log_map_wasserstein(X, Cref):
+def log_map_wasserstein(X, Cref, **kwargs):
     r"""Project matrices in tangent space by Wasserstein logarithmic map.
 
     The projection of a matrix :math:`\mathbf{X}` from SPD/HPD manifold
@@ -535,7 +539,7 @@ log_map_functions = {
 }
 
 
-def log_map(X, Cref, *, metric="riemann"):
+def log_map(X, Cref, *, metric="riemann", **kwargs):
     """Project matrices in tangent space by logarithmic map.
 
     Parameters
@@ -548,6 +552,10 @@ def log_map(X, Cref, *, metric="riemann"):
         Metric used for logarithmic map, can be:
         "euclid", "logchol", "logeuclid", "riemann", "wasserstein",
         or a callable function.
+    **kwargs : dict
+        The keyword arguments passed to the sub function.
+
+        .. versionadded:: 0.12
 
     Returns
     -------
@@ -569,7 +577,7 @@ def log_map(X, Cref, *, metric="riemann"):
     log_map_wasserstein
     """
     log_map_function = check_function(metric, log_map_functions)
-    return log_map_function(X, Cref)
+    return log_map_function(X, Cref, **kwargs)
 
 
 ###############################################################################
@@ -767,8 +775,72 @@ def innerproduct_euclid(X, Y, *args):
     """
     if Y is None:
         Y = X
-    G = _apply_inner_product(X, Y)
-    return G
+    return _apply_inner_product(X, Y)
+
+
+def innerproduct_logchol(X, Y, Cref):
+    r"""Log-Cholesky inner product.
+
+    Log-Cholesky inner product :math:`\mathbf{g}` between
+    symmetric matrices in tangent space :math:`\mathbf{X}`
+    and :math:`\mathbf{Y}` at :math:`\mathbf{C}_\text{ref}` is given in [1]_.
+
+    Parameters
+    ----------
+    X : ndarray, shape (..., n, n)
+        First symmetric matrices in tangent space at Cref.
+    Y : ndarray, shape (..., n, n) | None
+        Second symmetric matrices in tangent space at Cref.
+        If None, Y is set to X, giving the squared norm of X.
+    Cref : ndarray, shape (n, n)
+        Reference SPD matrix.
+
+    Returns
+    -------
+    G : float or ndarray, shape (...,)
+        Log-Cholesky inner product between X and Y.
+
+    Notes
+    -----
+    .. versionadded:: 0.12
+
+    See Also
+    --------
+    innerproduct
+
+    References
+    ----------
+    .. [1] `Riemannian geometry of symmetric positive definite matrices via
+        Cholesky decomposition
+        <https://arxiv.org/pdf/1908.09326>`_
+        Z. Lin. SIAM J Matrix Anal Appl, 2019, 40(4), pp. 1353-1370.
+    """
+    xp = check_matrix_pair(X, Cref, require_square=True)
+    C_chol = xp.linalg.cholesky(Cref)
+    eye_n = xp.eye(Cref.shape[-1], dtype=Cref.dtype, device=xpd(Cref))
+    C_invchol = xp.linalg.solve(C_chol, eye_n)
+
+    tri0, tri1 = tril_indices(X.shape[-1], -1)
+    diag0, diag1 = diag_indices(X.shape[-1])
+
+    def _inv_diff(W, L, Linv):
+        """Prop 4, Section 3.2 in [1]"""
+        S = Linv @ W @ ctranspose(Linv)
+        S12 = xp.zeros_like(S)
+        S12[..., tri0, tri1] = S[..., tri0, tri1]
+        S12[..., diag0, diag1] = S[..., diag0, diag1] / 2
+        dL = L @ S12
+        return dL[..., tri0, tri1], S12[..., diag0, diag1]
+
+    triX, diagX = _inv_diff(X, C_chol, C_invchol)
+    if Y is None:
+        triY, diagY = (triX, diagX)
+    else:
+        triY, diagY = _inv_diff(Y, C_chol, C_invchol)
+
+    tri = xp.sum(triX.conj() * triY, axis=-1)
+    diag = xp.sum(diagX.conj() * diagY, axis=-1)
+    return (tri + diag).real
 
 
 def innerproduct_logeuclid(X, Y, Cref):
@@ -818,10 +890,9 @@ def innerproduct_logeuclid(X, Y, Cref):
         V. Arsigny, P. Fillard, X. Pennec, N. Ayache.
         SIAM J Matrix Anal Appl, 2007, 29 (1), pp. 328-347
     """
-    if Y is None:
-        Y = X
-    G = _apply_inner_product(ddlogm(X, Cref), ddlogm(Y, Cref))
-    return G
+    X_ = ddlogm(X, Cref)
+    Y_ = X_ if Y is None else ddlogm(Y, Cref)
+    return _apply_inner_product(X_, Y_)
 
 
 def innerproduct_riemann(X, Y, Cref):
@@ -872,11 +943,10 @@ def innerproduct_riemann(X, Y, Cref):
         W. Förstner & B. Moonen.
         Geodesy-the Challenge of the 3rd Millennium, 2003
     """
-    if Y is None:
-        Y = X
     Cm12 = invsqrtm(Cref)
-    G = _apply_inner_product(Cm12 @ X @ Cm12, Cm12 @ Y @ Cm12)
-    return G
+    X_ = Cm12 @ X @ Cm12
+    Y_ = X_ if Y is None else Cm12 @ Y @ Cm12
+    return _apply_inner_product(X_, Y_)
 
 
 def _apply_inner_product(X, Y):
@@ -892,6 +962,7 @@ def _apply_inner_product(X, Y):
 
 innerproduct_functions = {
     "euclid": innerproduct_euclid,
+    "logchol": innerproduct_logchol,
     "logeuclid": innerproduct_logeuclid,
     "riemann": innerproduct_riemann,
 }
@@ -915,7 +986,7 @@ def innerproduct(X, Y, Cref, metric="riemann"):
         Reference matrix.
     metric : string | callable, default="riemann"
         Metric used for inner product, can be:
-        "euclid", "logeuclid", "riemann", or a callable function.
+        "euclid", "logchol", "logeuclid", "riemann", or a callable function.
 
     Returns
     -------
@@ -931,12 +1002,12 @@ def innerproduct(X, Y, Cref, metric="riemann"):
     See Also
     --------
     innerproduct_euclid
+    innerproduct_logchol
     innerproduct_logeuclid
     innerproduct_riemann
     """
     innerproduct_function = check_function(metric, innerproduct_functions)
-    G = innerproduct_function(X, Y, Cref)
-    return G
+    return innerproduct_function(X, Y, Cref)
 
 
 def norm(X, Cref, metric="riemann"):